chore: 네트워크 내부 파일 재배치
user, designer -> user
token관련 -> auth
