+++ GAM/GAM/Network/DTO/User/Request/ScrapDesignerRequestDTO.swift
