+++ GAM/GAM/Network/DTO/User/Request/UploadImageRequestDTO.swift
