+++ GAM/GAM/Network/DTO/User/Response/GetBrowseDesignerResponseDTO.swift
