+++ GAM/GAM/Network/DTO/User/Response/GetImageUrlResponseDTO.swift
@@ -1,5 +1,5 @@
 //
-//  ImageUrlResponseDTO.swift
+//  GetImageUrlResponseDTO.swift
 //  GAM
 //
 //  Created by Juhyeon Byun on 1/9/24.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-struct ImageUrlResponseDTO: Decodable {
+struct GetImageUrlResponseDTO: Decodable {
     let preSignedUrl: String
     let fileName: String
 }
