+++ GAM/GAM/Network/DTO/User/Response/GetScrapDesignerResponseDTO.swift
