+++ GAM/GAM/Network/DTO/User/Response/GetUserPortfolioResponseDTO.swift
