+++ GAM/GAM/Network/DTO/User/Response/PopularDesignerResponseDTO.swift
