+++ GAM/GAM/Network/DTO/User/Response/ScrapDesignerResponseDTO.swift
