+++ GAM/GAM/Network/DTO/User/Response/SearchDesignerResponseDTO.swift
