+++ GAM/GAM/Network/DTO/User/Response/SetDefaultMyPortfolioResponseDTO.swift
@@ -1,5 +1,5 @@
 //
-//  DefaultPortfolioResponseDTO.swift
+//  SetDefaultMyPortfolioResponseDTO.swift
 //  GAM
 //
 //  Created by Juhyeon Byun on 1/8/24.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-struct DefaultPortfolioResponseDTO: Decodable {
+struct SetDefaultMyPortfolioResponseDTO: Decodable {
     let workID: Int
 
     enum CodingKeys: String, CodingKey {
